requirements.txt
# Base Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0.0

# Authentication & JWT
djangorestframework-simplejwt>=5.2.0

# Database
psycopg2-binary>=2.9.0

# API Documentation
drf-spectacular>=0.26.0

# File handling
Pillow>=10.0.0

# Environment variables
python-dotenv>=1.0.0

# Development utilities
django-extensions>=3.2.0

# Production server
gunicorn>=21.0.0

# Production utilities
dj-database-url>=2.0.0
whitenoise>=6.0.0